add missing cast
git-svn-id: https://svn.r-project.org/R/trunk@73820 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/serialize.c
@@ -1537,7 +1537,7 @@ static int TryConvertString(void *obj, const char *inp, size_t inplen,
 {
     if (Riconv(obj, NULL, NULL, &buf, bufleft) == -1)
 	return -1;
-    return Riconv(obj, &inp, &inplen, &buf, bufleft);
+    return (int) Riconv(obj, &inp, &inplen, &buf, bufleft);
 }
 
 static SEXP
